Use arbirtary-sized integers when working with plan costs
The numbers get pretty big when computing the cost of a search plan,
which corresponds to the number of assignments of the free variables the
plan implies.

For order n matricies, the rank is (2n + 1), so the number of free
variables is (n^2 - (2n + 1)). For n = 4, the number of free variables
is 7. For n = 5, it's 14. For n = 6, it's 23.

For the absolute most naive search plans with p free variables that
consider all possible assignments of the free variables, the cost is

((n^2)!)/((n^2 - p)!)

For n = 4, the cost is 57,657,600.
For n = 5, the cost is 3.885885152e17.
For n = 6, the cost is 5.973857142e31.

The max signed 64-bit integer is (2^63 - 1) = 4.611686018e18